+++ src/core/Buttons/Button/Button.tsx
@@ -79,9 +79,9 @@ export const Button: ButtonComponent = React.forwardRef(
         ref={ref}
         className={cx(
           'iui-button',
+          `iui-${styleType}`,
           {
             [`iui-${size}`]: !!size,
-            [`iui-${styleType}`]: styleType !== 'default',
           },
           className,
         )}
@@ -91,14 +91,14 @@ export const Button: ButtonComponent = React.forwardRef(
       >
         {startIcon &&
           React.cloneElement(startIcon, {
-            className: cx('iui-icon', startIcon.props.className),
+            className: cx('iui-button-icon', startIcon.props.className),
           })}
 
-        {children && <span className='iui-label'>{children}</span>}
+        {children && <span className='iui-button-label'>{children}</span>}
 
         {endIcon &&
           React.cloneElement(endIcon, {
-            className: cx('iui-icon', endIcon.props.className),
+            className: cx('iui-button-icon', endIcon.props.className),
           })}
       </Element>
     );